Fixed up the formatting of dates for historical data, might update action next to preformat and allow for more date options.
