+++ src/utils/convertHistoricalData.ts
@@ -1,20 +1,8 @@
 import { HistoricalPriceDataResponse } from "@/lib/types/HistoricalPriceDataResponse";
+import { reshapeHistoricalData } from "./reshapeHistoricalData";
 
 export const convertHistoricalData = (data: HistoricalPriceDataResponse) => {
-  const reorganizedData = {
-    prices: data.prices,
-    volumes: data.total_volumes,
-  };
+  const reshapedData = reshapeHistoricalData(data);
 
-  const reshapedData = reorganizedData.prices.map((price, index) => {
-    if (index < 24) {
-      const volume = Math.abs(
-        reorganizedData.volumes[index][1] -
-          reorganizedData.volumes[index + 24][1]
-      );
-      return { price: price[1], volume: volume, date: new Date(price[0]) };
-    }
-  });
-
-  return reshapedData.filter((data) => data !== undefined);
+  return reshapedData;
 };
