+++ src/utils/formatChartDate.ts
@@ -1,0 +1,7 @@
+export const formatChartDate = (timeStamp: number) => {
+  const dateObj = new Date(timeStamp);
+  const dateStr =
+    dateObj.toLocaleDateString() + ", " + dateObj.toLocaleTimeString();
+
+  return dateStr;
+};
